requirements.txt
ultralytics>=8.2.0
opencv-python>=4.9.0.80
Pillow>=10.3.0
numpy>=1.24.3
customtkinter>=5.2.2
matplotlib>=3.7.1
pandas>=2.0.3
